test(my-scenes): split my-scenes specs and add module-local fixtures
